Moved Dictionary to dataflow/common; Refactored code
